+++ BackEnd/src/main/java/com/example/demo/model/MenuItem.java
@@ -1,6 +1,7 @@
 package com.example.demo.model;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Inheritance(strategy = InheritanceType.SINGLE_TABLE)
@@ -12,7 +13,7 @@ public class MenuItem extends Item {
 
     }
 
-    public MenuItem(String name,  Double price, String description,String imageURL){
+    public MenuItem(String name, BigDecimal price, String description, String imageURL){
         super(name, price, description, imageURL);
     }
